modules/device/requirements.txt
azure-iot-device~=2.7.0
psutil
py-cpuinfo